requirements.txt
django>=5.2.6
djangorestframework
whitenoise
gunicorn
dj-database-url
python-dotenv
psycopg2-binary
django-cors-headers